requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
EbookLib==0.19
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
idna==3.10
jieba3k==0.35.1
joblib==1.5.1
lxml==5.4.0
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.2
readability-lxml==0.8.4.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
